fin de seccion 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -9,13 +10,13 @@ import { MessageService } from '../_services/message.service';
   styleUrls: ['./messages.component.css']
 })
 export class MessagesComponent implements OnInit {
-  messages: Message[]=[];
+  messages: Message[] = [];
   pagination: Pagination;
-  container: string= "Unread";
+  container: string = "Unread";
   pageNumber = 1;
   pageSize = 5;
-  loading= false;
-  constructor(private messageService: MessageService) { }
+  loading = false;
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -24,19 +25,27 @@ export class MessagesComponent implements OnInit {
 
   loadMessages() {//el container pasa como undefined desde la prop no adquiere valor investigar despues
     //solu le he delarado del tipo que es string
-    this.loading=true;
+    this.loading = true;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe
       (response => {
         this.messages = response.result;
         this.pagination = response.pagination;
-        this.loading=false;
+        this.loading = false;
       })
   }
 
-  deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex(m => m.id===id),1);
-    })
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm delete message', 'This can not be undone')
+      .subscribe(result => {
+        if (result) {
+          
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          })
+
+        }
+      })
+
   }
 
   pageChanged(event: any) {
